Stupid typo (nw)

+++ src/devices/cpu/m68000/m68kcpu.cpp
@@ -1322,7 +1322,7 @@ void m68000_base_device::init32mmu(address_space &space, address_space &ospace)
 				if (m_mmu_tmp_buserror_occurred)
 					return ~0;
 				if (WORD_ALIGNED(address)) { // 2
-					u16 result = m_space->read_word(address0) << 16;
+					u32 result = m_space->read_word(address0) << 16;
 					return result | m_space->read_word(address2);
 				}
 				u32 address1 = pmmu_translate_addr(address+1);
